+++ libs/qt/release_notes.md
@@ -2,6 +2,7 @@
 
 ## upcoming
 
+* [changed] StatusMaterialIconWidget clickable.
 * [new] Editable label widget.